Error Fixed , GenRepo not done
