+++ DeviceManagement_WebApp/Repository/DeviceRepository.cs
@@ -14,7 +14,6 @@ namespace DeviceManagement_WebApp.Repository
         }
 
 
-
     }
